fix msvc build.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@6784 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_core_file.c
@@ -130,7 +130,7 @@ SWITCH_DECLARE(switch_status_t) switch_core_file_read(switch_file_handle_t *fh, 
 									   fh->native_rate,
 									   orig_len * 10,
 									   fh->samplerate,
-									   orig_len * 10,
+									   (uint32_t)orig_len * 10,
 									   fh->memory_pool) != SWITCH_STATUS_SUCCESS) {
 				switch_log_printf(SWITCH_CHANNEL_LOG, SWITCH_LOG_CRIT, "Unable to create resampler!\n");
 				return SWITCH_STATUS_GENERR;
@@ -180,7 +180,7 @@ SWITCH_DECLARE(switch_status_t) switch_core_file_write(switch_file_handle_t *fh,
 									   fh->native_rate,
 									   orig_len * 10,
 									   fh->samplerate,
-									   orig_len * 10,
+									   (uint32_t)orig_len * 10,
 									   fh->memory_pool) != SWITCH_STATUS_SUCCESS) {
 				switch_log_printf(SWITCH_CHANNEL_LOG, SWITCH_LOG_CRIT, "Unable to create resampler!\n");
 				return SWITCH_STATUS_GENERR;
